+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Unreleased
 
 - Fix [#792](https://github.com/squint-cljs/squint/issues/792): `prn` `js/undefined` as `nil`
+- Fix [#793](https://github.com/squint-cljs/squint/issues/793): fix `yield*` IIFE (needs wrapping in parens)
 
 ## v0.9.183